Fix id/ancestor_id dtype preservation across multiple functions
- Apply @preserve_id_dtypes decorator to add_global_root,
  add_inner_knuckles_asexual, add_inner_niblings_asexual, and
  sample_triplet_comparisons_asexual to prevent uint64->float64 casts
- Fix coarsen_mask to use original id dtype instead of hardcoded int
- Cast numba array indices to np.intp in unfurl_traversal_semiorder and
  find_pair_mrca_id to fix uint64 array indexing TypingError
- Make assert_dtype_consistency robust: handle no-arg functions, skip
  uncastable data, suppress warnings during dtype checks, catch
  exceptions from functions that fail on cast data

+++ phyloframe/legacy/_alifestd_find_pair_mrca_id_asexual.py
@@ -107,10 +107,10 @@ def alifestd_find_pair_mrca_id_asexual(
     ):
         raise NotImplementedError("non-contiguous ids not yet supported")
 
-    ancestor_ids = phylogeny_df["ancestor_id"].to_numpy()
+    ancestor_ids = phylogeny_df["ancestor_id"].to_numpy(dtype=np.intp)
     result = _alifestd_find_pair_mrca_id_asexual_fast_path(
         ancestor_ids,
-        first,
-        second,
+        int(first),
+        int(second),
     )
     return None if result == -1 else int(result)
